+++ Src/Private/SharedUtilsFunctions.ps1
@@ -850,3 +850,121 @@ function Get-VBRDebugObject {
         Name = "ObjectRepositor1", "ObjectRepositor2", "ObjectRepositor3", "ObjectRepositor4", "ObjectRepositor5", "ObjectRepositor6", "ObjectRepositor7"
     }
 }
+
+function New-VBRConnection {
+    <#
+    .SYNOPSIS
+        Uses New-VBRConnection to store the connection in a global parameter
+    .DESCRIPTION
+        Creates a Veeam Server connection and stores it in global variable $Global:DefaultVeeamBR.
+        An FQDN or IP, credentials, and ignore certificate boolean
+    .OUTPUTS
+        Returns the Veeam Server connection.
+    .EXAMPLE
+    New-VBRConnection -Endpoint <FQDN or IP> -Port <default 9419> -Credential $(Get-Credential)
+
+    #>
+
+    [CmdletBinding()]
+    Param(
+
+        [Parameter(Position=0,mandatory=$true)]
+        [string]$Endpoint,
+
+        [Parameter(Position=1,mandatory=$true)]
+        [string]$Port,
+
+        [Parameter(Mandatory=$true,ParameterSetName="Credential")]
+        [ValidateNotNullOrEmpty()]
+        [Management.Automation.PSCredential]$Credential
+
+    )
+
+    $apiUrl = "https://$($Endpoint):$($Port)/api/oauth2/token"
+
+    $User = $Credential.UserName
+    $Pass = $Credential.GetNetworkCredential().Password
+
+    # Define the headers for the API request
+    $headers = @{
+        "Content-Type"  = "application/x-www-form-urlencoded"
+        "x-api-version" = "1.1-rev0"
+    }
+
+    ## TO-DO: Grant_type options
+    $body = @{
+        "grant_type" = "password"
+        "username"   = $User
+        "password"   = $Pass
+    }
+
+    # Send an authentication request to obtain a session token
+    try {
+        $response = Invoke-RestMethod -Uri $apiUrl -Headers $headers -Method Post -Body $body -SkipCertificateCheck
+
+        if (($response.access_token) -or ($response.StatusCode -eq 200) ) {
+            Write-Output "Successfully authenticated."
+            $VBRAuthentication = [PSCustomObject]@{
+                Session_endpoint       = $Endpoint
+                Session_port           = $Port
+                Session_access_token  = $response.access_token
+            }
+
+            return $VBRAuthentication
+        }
+        else {
+            Write-Output "Authentication failed. Status code: $($response.StatusCode), Message: $($response.Content)"
+        }
+    }
+    catch {
+        Write-Output "An error occurred: $($_.Exception.Message)"
+    }
+}
+
+# Variable translating Icon to Image Path ($IconPath)
+$script:Images = @{
+    "VBR_Server" = "VBR_server.png"
+    "VBR_Repository" = "VBR_Repository.png"
+    "VBR_NAS" = "NAS.png"
+    "VBR_Deduplicating_Storage" = "Deduplication.png"
+    "VBR_Linux_Repository" = "Linux_Repository.png"
+    "VBR_Windows_Repository" = "Windows_Repository.png"
+    "VBR_Cloud_Repository" = "Cloud_Repository.png"
+    "VBR_Object_Repository" = "Object_Storage.png"
+    "VBR_Object" = "Object_Storage_support.png"
+    "VBR_Amazon_S3_Compatible" = "S3-compatible.png"
+    "VBR_Amazon_S3" = "AWS S3.png"
+    "VBR_Azure_Blob" = "Azure Blob.png"
+    "VBR_Server_DB" = "Microsoft_SQL_DB.png"
+    "VBR_Proxy" = "Veeam_Proxy.png"
+    "VBR_Proxy_Server" = "Proxy_Server.png"
+    "VBR_Wan_Accel" = "WAN_accelerator.png"
+    "VBR_SOBR" = "Logo_SOBR.png"
+    "VBR_SOBR_Repo" = "Scale_out_Backup_Repository.png"
+    "VBR_LOGO" = "Veeam_logo.png"
+    "VBR_No_Icon" = "no_icon.png"
+    'VBR_Storage_NetApp' = "Storage_NetApp.png"
+    'VBR_vCenter_Server' = 'vCenter_server.png'
+    'VBR_ESXi_Server' = 'ESXi_host.png'
+    'VBR_HyperV_Server' = 'Hyper-V_host.png'
+    'VBR_Server_EM' = 'Veeam_Backup_Enterprise_Manager.png'
+    'VBR_Tape_Server' = 'Tape_Server.png'
+    'VBR_Tape_Library' = 'Tape_Library.png'
+    'VBR_Tape_Drive' = 'Tape_Drive.png'
+    'VBR_Tape_Vaults' = 'Tape encrypted.png'
+    "VBR_Server_DB_PG" = "PostGre_SQL_DB.png"
+    "VBR_LOGO_Footer" = "verified_recoverability.png"
+    "VBR_AGENT_Container" = "Folder.png"
+    "VBR_AGENT_AD" = "Server.png"
+    "VBR_AGENT_MC" = "Task list.png"
+    "VBR_AGENT_IC" = "Workstation.png"
+    "VBR_AGENT_CSV" = "CSV_Computers.png"
+    "VBR_AGENT_AD_Logo" = "Microsoft Active Directory.png"
+    "VBR_AGENT_CSV_Logo" = "File.png"
+    "VBR_AGENT_Server" = "Server_with_Veeam_Agent.png"
+    "VBR_vSphere" = "VMware_vSphere.png"
+    "VBR_HyperV" = "Microsoft_SCVMM.png"
+    "VBR_Tape" = "Tape.png"
+    "VBR_Service_Providers" = "Veeam_Service_Provider_Console.png"
+    "VBR_Service_Providers_Server" = "Veeam_Service_Provider_Server.png"
+}