+++ CHANGELOG.md
@@ -7,6 +7,21 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ##### This project is community maintained and has no sponsorship from Veeam, its employees or any of its affiliates.
 
+## [0.8.10] - 2024-09-12
+
+### Changed
+
+- Migrate infrastructure diagram to Veeam.Diagrammer
+
+### Changed
+
+- Increase Diagrammer.Core minimum requirement to v0.2.7
+- Increase Veeam.Diagrammer minimum requirement to v0.6.6
+
+### Remove
+
+- Remove unused icons folder and images
+
 ## [0.8.9] - 2024-08-31
 
 ### Added